feat: did some refactor
