+++ pkgs/prism/config_test.go
@@ -1,4 +1,4 @@
-package routes
+package prism
 
 import "testing"
